Support referencing any contents in the org file
This is achieved by calling `org-export-to-buffer' to export a field (which is
an org subtree) instead of treating the field as a string and calling
`org-export-string-as' to export this string. The problem with the later (old)
approach is that the HTML export backend is only able to see the current
field (org subtree), and therefore the current field is not able to reference
any contents outside of the field.